+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 * When in private repo without `DEMSITO_SDK_GITHUB_TOKEN` configured, get_remote_file will take files from the local origin/master.
 * Enhanced the **unify** command when giving input of a file and not a directory return a clear error message.
+* Added a validation to ensure integrations are not skipped and at least one test playbook is not skipped for each integration or script.
 * Added new validation for the `xsoar_config.json` file in the **validate** command.
 
 # 1.4.0